Colorクラスno
