+++ source/block_bingo/BlockBingoCoordinate.py
@@ -5,18 +5,7 @@
 """
 from enum import Enum, auto
 import numpy as np
-
-class Color(Enum):
-    """
-    ブロック、ブロックサークルまたは交点サークルの色。
-    """
-    NONE = auto()   # ブロックが置かれていない状態を表現するために用いる
-    RED = auto()
-    BLUE = auto()
-    YELLOW = auto()
-    GREEN = auto()
-    BLACK = auto()
-    WHITE = auto()
+from source.Color import Color
 
 class BlockCirclesCoordinate():
     """
